chore: rearrange settings
